Fix compilation on 32-bit platform
fixes #295

+++ values.go
@@ -74,7 +74,7 @@ func convertSimpleArgument(ci *pgtype.ConnInfo, arg interface{}) (interface{}, e
 		}
 		return int64(arg), nil
 	case uint:
-		if arg > math.MaxInt64 {
+		if uint64(arg) > math.MaxInt64 {
 			return nil, errors.Errorf("arg too big for int64: %v", arg)
 		}
 		return int64(arg), nil
